fix inbox apis authentication + refactoring
