+++ components/api/api-security/src/main/java/org/eclipse/dirigible/components/api/security/UserFacade.java
@@ -9,27 +9,26 @@
  */
 package org.eclipse.dirigible.components.api.security;
 
-import static java.text.MessageFormat.format;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Locale;
-import java.util.stream.Collectors;
-
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.websocket.Session;
-
 import org.eclipse.dirigible.commons.config.Configuration;
 import org.eclipse.dirigible.components.api.http.HttpRequestFacade;
 import org.eclipse.dirigible.components.api.http.HttpSessionFacade;
 import org.eclipse.dirigible.components.base.context.ContextException;
 import org.eclipse.dirigible.components.base.context.ThreadContextFacade;
+import org.eclipse.dirigible.components.base.util.AuthoritiesUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Locale;
+
+import static java.text.MessageFormat.format;
 
 /**
  * The Class UserFacade.
@@ -71,6 +70,52 @@ public class UserFacade {
     private static final String AUTH = "authorization";
 
     /**
+     * Checks if the user is in role.
+     *
+     * @param role the role
+     * @return true, if the user is in role
+     */
+    public static final boolean isInRole(String role) {
+        if (Configuration.isAnonymousModeEnabled() || Configuration.isAnonymousUserEnabled()) {
+            return true;
+        }
+        try {
+            return HttpRequestFacade.isUserInRole(role);
+        } catch (Exception e) {
+            if (logger.isErrorEnabled()) {
+                logger.error(e.getMessage());
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Gets the user name by a given request as parameter.
+     *
+     * @param request the request
+     * @return the user name
+     */
+    public static final String getName(HttpServletRequest request) {
+        if (request == null) {
+            return getName();
+        }
+        // HTTP case
+        String userName = null;
+        try {
+            userName = request.getRemoteUser();
+        } catch (Exception e) {
+            if (logger.isErrorEnabled()) {
+                logger.error(e.getMessage());
+            }
+        }
+        if (userName != null) {
+            return userName;
+        }
+        return getName();
+    }
+
+    /**
      * Gets the user name.
      *
      * @return the user name
@@ -90,24 +135,99 @@ public class UserFacade {
     }
 
     /**
-     * Checks if the user is in role.
+     * Gets the remote user.
      *
-     * @param role the role
-     * @return true, if the user is in role
+     * @return the remote user
      */
-    public static final boolean isInRole(String role) {
-        if (Configuration.isAnonymousModeEnabled() || Configuration.isAnonymousUserEnabled()) {
-            return true;
-        }
+    private static String getRemoteUser() {
         try {
-            return HttpRequestFacade.isUserInRole(role);
+            if (HttpRequestFacade.isValid()) {
+                return HttpRequestFacade.getRemoteUser();
+            }
         } catch (Exception e) {
             if (logger.isErrorEnabled()) {
                 logger.error(e.getMessage());
             }
         }
+        return null;
+    }
 
-        return false;
+    /**
+     * Gets the anonymous user.
+     *
+     * @return the anonymous user
+     */
+    private static String getAnonymousUser() {
+        String userName = null;
+        if (Configuration.isAnonymousModeEnabled()) {
+            try {
+                userName = getContextProperty(DIRIGIBLE_ANONYMOUS_IDENTIFIER);
+            } catch (ContextException e) {
+                if (logger.isErrorEnabled()) {
+                    logger.error(e.getMessage());
+                }
+            }
+        } else if (Configuration.isAnonymousUserEnabled()) {
+            try {
+                userName = getContextProperty(DIRIGIBLE_ANONYMOUS_USER);
+                if (userName == null) {
+                    userName = setAnonymousUser();
+                }
+            } catch (ContextException e) {
+                if (logger.isErrorEnabled()) {
+                    logger.error(e.getMessage());
+                }
+            }
+        } else if (Configuration.isJwtModeEnabled()) {
+            try {
+                userName = getContextProperty(DIRIGIBLE_JWT_USER);
+            } catch (ContextException e) {
+                if (logger.isErrorEnabled()) {
+                    logger.error(e.getMessage());
+                }
+            }
+        }
+        return userName;
+    }
+
+    /**
+     * Gets the context property.
+     *
+     * @param property the property
+     * @return the context property
+     * @throws ContextException the context exception
+     */
+    private static String getContextProperty(String property) throws ContextException {
+        if (HttpSessionFacade.isValid()) {
+            return HttpSessionFacade.getAttribute(property);
+        } else if (ThreadContextFacade.isValid()) {
+            Object value = ThreadContextFacade.get(property);
+            if ((value != null) && (value instanceof String)) {
+                return (String) value;
+            }
+        }
+        return null;
+    }
+
+    /**
+     * Sets the anonymous user.
+     *
+     * @return the string
+     */
+    private static String setAnonymousUser() {
+        String userName = null;
+        String anonymousUserNamePropertyName = Configuration.get(DIRIGIBLE_ANONYMOUS_USER_NAME_PROPERTY_NAME);
+        if (anonymousUserNamePropertyName != null) {
+            userName = Configuration.get(anonymousUserNamePropertyName);
+            try {
+                setName(userName);
+            } catch (ContextException e) {
+                if (logger.isInfoEnabled()) {
+                    logger.info("Error while setting userName from DIRIGIBLE_ANONYMOUS_USER_PROPERTY_NAME.", e);
+                }
+            }
+        }
+        return userName;
     }
 
     /**
@@ -140,28 +260,20 @@ public class UserFacade {
     }
 
     /**
-     * Gets the user name by a given request as parameter.
+     * Sets the context property.
      *
-     * @param request the request
-     * @return the user name
+     * @param property the property
+     * @param value the value
+     * @throws ContextException the context exception
      */
-    public static final String getName(HttpServletRequest request) {
-        if (request == null) {
-            return getName();
-        }
-        // HTTP case
-        String userName = null;
-        try {
-            userName = request.getRemoteUser();
-        } catch (Exception e) {
-            if (logger.isErrorEnabled()) {
-                logger.error(e.getMessage());
+    private static void setContextProperty(String property, String value) throws ContextException {
+        if (HttpSessionFacade.isValid()) {
+            HttpSessionFacade.setAttribute(property, value);
+        } else {
+            if (ThreadContextFacade.isValid()) {
+                ThreadContextFacade.set(property, value);
             }
         }
-        if (userName != null) {
-            return userName;
-        }
-        return getName();
     }
 
     /**
@@ -284,121 +396,6 @@ public class UserFacade {
         Authentication authentication = SecurityContextHolder.getContext()
                                                              .getAuthentication();
         Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-        return authorities.stream()
-                          .map(GrantedAuthority::getAuthority)
-                          .collect(Collectors.toList());
-    }
-
-    /**
-     * Gets the context property.
-     *
-     * @param property the property
-     * @return the context property
-     * @throws ContextException the context exception
-     */
-    private static String getContextProperty(String property) throws ContextException {
-        if (HttpSessionFacade.isValid()) {
-            return HttpSessionFacade.getAttribute(property);
-        } else if (ThreadContextFacade.isValid()) {
-            Object value = ThreadContextFacade.get(property);
-            if ((value != null) && (value instanceof String)) {
-                return (String) value;
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Sets the context property.
-     *
-     * @param property the property
-     * @param value the value
-     * @throws ContextException the context exception
-     */
-    private static void setContextProperty(String property, String value) throws ContextException {
-        if (HttpSessionFacade.isValid()) {
-            HttpSessionFacade.setAttribute(property, value);
-        } else {
-            if (ThreadContextFacade.isValid()) {
-                ThreadContextFacade.set(property, value);
-            }
-        }
-    }
-
-    /**
-     * Gets the remote user.
-     *
-     * @return the remote user
-     */
-    private static String getRemoteUser() {
-        try {
-            if (HttpRequestFacade.isValid()) {
-                return HttpRequestFacade.getRemoteUser();
-            }
-        } catch (Exception e) {
-            if (logger.isErrorEnabled()) {
-                logger.error(e.getMessage());
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Gets the anonymous user.
-     *
-     * @return the anonymous user
-     */
-    private static String getAnonymousUser() {
-        String userName = null;
-        if (Configuration.isAnonymousModeEnabled()) {
-            try {
-                userName = getContextProperty(DIRIGIBLE_ANONYMOUS_IDENTIFIER);
-            } catch (ContextException e) {
-                if (logger.isErrorEnabled()) {
-                    logger.error(e.getMessage());
-                }
-            }
-        } else if (Configuration.isAnonymousUserEnabled()) {
-            try {
-                userName = getContextProperty(DIRIGIBLE_ANONYMOUS_USER);
-                if (userName == null) {
-                    userName = setAnonymousUser();
-                }
-            } catch (ContextException e) {
-                if (logger.isErrorEnabled()) {
-                    logger.error(e.getMessage());
-                }
-            }
-        } else if (Configuration.isJwtModeEnabled()) {
-            try {
-                userName = getContextProperty(DIRIGIBLE_JWT_USER);
-            } catch (ContextException e) {
-                if (logger.isErrorEnabled()) {
-                    logger.error(e.getMessage());
-                }
-            }
-        }
-        return userName;
-    }
-
-    /**
-     * Sets the anonymous user.
-     *
-     * @return the string
-     */
-    private static String setAnonymousUser() {
-        String userName = null;
-        String anonymousUserNamePropertyName = Configuration.get(DIRIGIBLE_ANONYMOUS_USER_NAME_PROPERTY_NAME);
-        if (anonymousUserNamePropertyName != null) {
-            userName = Configuration.get(anonymousUserNamePropertyName);
-            try {
-                setName(userName);
-            } catch (ContextException e) {
-                if (logger.isInfoEnabled()) {
-                    logger.info("Error while setting userName from DIRIGIBLE_ANONYMOUS_USER_PROPERTY_NAME.", e);
-                }
-            }
-        }
-        return userName;
+        return AuthoritiesUtil.toRoleNames(authorities);
     }
 }
