+++ components/core/core-tenants/src/main/java/org/eclipse/dirigible/components/tenants/security/CustomUserDetailsService.java
@@ -11,6 +11,7 @@ package org.eclipse.dirigible.components.tenants.security;
 
 import org.eclipse.dirigible.components.base.tenant.Tenant;
 import org.eclipse.dirigible.components.base.tenant.TenantContext;
+import org.eclipse.dirigible.components.base.util.AuthoritiesUtil;
 import org.eclipse.dirigible.components.tenants.domain.User;
 import org.eclipse.dirigible.components.tenants.service.UserService;
 import org.slf4j.Logger;
