+++ components/security/security-keycloak/src/main/java/org/eclipse/dirigible/components/security/keycloak/CustomOidcUserService.java
@@ -12,8 +12,8 @@ package org.eclipse.dirigible.components.security.keycloak;
 import org.eclipse.dirigible.commons.config.DirigibleConfig;
 import org.eclipse.dirigible.components.base.http.roles.Roles;
 import org.eclipse.dirigible.components.base.tenant.TenantContext;
+import org.eclipse.dirigible.components.base.util.AuthoritiesUtil;
 import org.eclipse.dirigible.components.tenants.domain.User;
-import org.eclipse.dirigible.components.tenants.security.AuthoritiesUtil;
 import org.eclipse.dirigible.components.tenants.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
